:recycle: abstract out remove btn and modal as their own components
The motivation here is to clearly separate the responsibilies for better composability and maintenance
